+++ sdk/security/keyvault/azcertificates/client_test.go
@@ -27,6 +27,7 @@ import (
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/to"
 	"github.com/Azure/azure-sdk-for-go/sdk/internal/mock"
 	"github.com/Azure/azure-sdk-for-go/sdk/internal/recording"
+	azcred "github.com/Azure/azure-sdk-for-go/sdk/internal/test/credential"
 	"github.com/Azure/azure-sdk-for-go/sdk/security/keyvault/azcertificates"
 	"github.com/stretchr/testify/require"
 )
@@ -285,7 +286,7 @@ func TestDisableChallengeResourceVerification(t *testing.T) {
 				},
 				DisableChallengeResourceVerification: test.disableVerify,
 			}
-			client, err := azcertificates.NewClient(vaultURL, &FakeCredential{}, options)
+			client, err := azcertificates.NewClient(vaultURL, &azcred.Fake{}, options)
 			require.NoError(t, err)
 			pager := client.NewListCertificatePropertiesPager(nil)
 			_, err = pager.NextPage(context.Background())