+++ sdk/security/keyvault/azcertificates/utils_test.go
@@ -16,16 +16,13 @@ import (
 	"time"
 
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
-	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
-	"github.com/Azure/azure-sdk-for-go/sdk/azidentity"
 	"github.com/Azure/azure-sdk-for-go/sdk/internal/recording"
+	azcred "github.com/Azure/azure-sdk-for-go/sdk/internal/test/credential"
 	"github.com/Azure/azure-sdk-for-go/sdk/security/keyvault/azcertificates"
 	"github.com/stretchr/testify/require"
 )
 
 const recordingDirectory = "sdk/security/keyvault/azcertificates/testdata"
-
-const fakeVaultURL = "https://test.vault.azure.net/"
 
 var (
 	certsToPurge = struct {
@@ -35,6 +32,8 @@ var (
 
 	credential azcore.TokenCredential
 	vaultURL   string
+
+	fakeVaultURL = fmt.Sprintf("https://%s.vault.azure.net/", recording.SanitizedValue)
 )
 
 func TestMain(m *testing.M) {
@@ -59,20 +58,13 @@ func run(m *testing.M) int {
 		}()
 	}
 
-	if recording.GetRecordMode() == recording.PlaybackMode {
-		credential = &FakeCredential{}
-	} else {
-		tenantId := getEnvVar("AZCERTIFICATES_TENANT_ID", "")
-		clientId := getEnvVar("AZCERTIFICATES_CLIENT_ID", "")
-		secret := getEnvVar("AZCERTIFICATES_CLIENT_SECRET", "")
-		var err error
-		credential, err = azidentity.NewClientSecretCredential(tenantId, clientId, secret, nil)
-		if err != nil {
-			panic(err)
-		}
+	var err error
+	credential, err = azcred.New(nil)
+	if err != nil {
+		panic(err)
 	}
 
-	vaultURL = getEnvVar("AZURE_KEYVAULT_URL", fakeVaultURL)
+	vaultURL = recording.GetEnvVariable("AZURE_KEYVAULT_URL", fakeVaultURL)
 
 	if recording.GetRecordMode() != recording.LiveMode {
 		err := recording.RemoveRegisteredSanitizers([]string{
@@ -134,33 +126,6 @@ func getName(t *testing.T, prefix string) string {
 	_, err := h.Write([]byte(t.Name()))
 	require.NoError(t, err)
 	return prefix + fmt.Sprint(h.Sum32())
-}
-
-func getEnvVar(envVar string, fakeValue string) string {
-	// get value
-	value := fakeValue
-	if recording.GetRecordMode() == recording.LiveMode || recording.GetRecordMode() == recording.RecordingMode {
-		value = os.Getenv(envVar)
-		if value == "" {
-			panic("no value for " + envVar)
-		}
-	}
-
-	// sanitize value
-	if fakeValue != "" && recording.GetRecordMode() == recording.RecordingMode {
-		err := recording.AddGeneralRegexSanitizer(fakeValue, value, nil)
-		if err != nil {
-			panic(err)
-		}
-	}
-
-	return value
-}
-
-type FakeCredential struct{}
-
-func (f *FakeCredential) GetToken(ctx context.Context, options policy.TokenRequestOptions) (azcore.AccessToken, error) {
-	return azcore.AccessToken{Token: "faketoken", ExpiresOn: time.Now().Add(time.Hour).UTC()}, nil
 }
 
 func cleanUpCert(t *testing.T, client *azcertificates.Client, name string) {