+++ sdk/storage/azfile/internal/shared/shared.go
@@ -127,10 +127,8 @@ func ParseConnectionString(connectionString string) (ParsedConnectionString, err
 				ServiceURL: fmt.Sprintf("%v?%v", fileEndpoint, sharedAccessSignature),
 			}, nil
 		}
-	} else {
-		if accountName == "" {
-			return ParsedConnectionString{}, errors.New("connection string missing AccountName")
-		}
+	} else if accountName == "" {
+		return ParsedConnectionString{}, errors.New("connection string missing AccountName")
 	}
 
 	protocol, ok := connStrMap["DefaultEndpointsProtocol"]